refactor(core): move abbreviatedLink into core
User Profile needs the same link formatting the You tab uses, but it can't
reach the helper: myaccount depends on :apps:flipcash:shared:menu, not
:apps:flipcash:features:menu. Both feature modules get :apps:flipcash:core
from the convention plugin, so the helper moves there and becomes public.

One implementation means the two screens can't drift on how a link reads.
